Revert "refactor(styled): split index.ts into seperate files"
This reverts commit 7a014f76b8b2a58abc80953173f0f524216ce1fc.
