--- packages/styled/src/builders/transition.ts
@@ -1,2 +1,0 @@
-/** @deprecated */
-export const TRANSITION_DURATION = 0.2
