--- packages/styled/src/context.ts
@@ -1,7 +1,0 @@
-export interface Context {
-  cancelHalfLeadingPx?: number
-  hasVerticalPadding?: boolean
-  boxShadowTransition?: boolean
-  colorTransition?: boolean
-  backgroundColorTransition?: boolean
-}
