+++ packages/styled/src/lib.ts
@@ -1,6 +1,4 @@
-import { EffectType, Key } from '@charcoal-ui/theme'
-import { notDisabledSelector, disabledSelector } from '@charcoal-ui/utils'
-import { CSSObject } from 'styled-components'
+import { Key } from '@charcoal-ui/theme'
 import { unreachable } from './util'
 
 /**
@@ -180,18 +178,3 @@ export type ModifiedArgumented<
 }
 
 export const variable = (value: string) => `var(${value})`
-
-export function isSupportedEffect(effect: Key): effect is EffectType {
-  return ['hover', 'press', 'disabled'].includes(effect as string)
-}
-
-export function onEffectPseudo(effect: EffectType, css: CSSObject) {
-  return effect === 'hover'
-    ? { '&:hover': { [notDisabledSelector]: css } }
-    : effect === 'press'
-    ? { '&:active': { [notDisabledSelector]: css } }
-    : // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition
-    effect === 'disabled'
-    ? { [disabledSelector]: css }
-    : unreachable(effect)
-}
